+++ src/vec3d/dist.c
@@ -6,14 +6,14 @@
 /*   By: dderny <dderny@student.42lyon.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/09 02:16:44 by dderny            #+#    #+#             */
-/*   Updated: 2025/05/11 14:43:58 by dderny           ###   ########.fr       */
+/*   Updated: 2025/05/24 05:41:54 by dderny           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "vec3d.h"
 #include <math.h>
 
-inline double	vec3d_dist(const t_vec3d a, const t_vec3d b)
+inline float	vec3d_dist(const t_vec3d a, const t_vec3d b)
 {
 	return (sqrt(vec3d_sqrlen(vec3d_sub(a, b))));
 }
