+++ src/vec3d/dot.c
@@ -6,13 +6,13 @@
 /*   By: dderny <dderny@student.42lyon.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/09 02:16:32 by dderny            #+#    #+#             */
-/*   Updated: 2025/05/11 14:43:58 by dderny           ###   ########.fr       */
+/*   Updated: 2025/05/24 05:41:54 by dderny           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "vec3d.h"
 
-inline double	vec3d_dot(const t_vec3d a, const t_vec3d b)
+inline float	vec3d_dot(const t_vec3d a, const t_vec3d b)
 {
 	return (a.x * b.x + a.y * b.y + a.z * b.z);
 }
