+++ src/vec3d/lerp.c
@@ -6,13 +6,13 @@
 /*   By: dderny <dderny@student.42lyon.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/09 02:18:26 by dderny            #+#    #+#             */
-/*   Updated: 2025/05/11 14:43:58 by dderny           ###   ########.fr       */
+/*   Updated: 2025/05/24 05:41:54 by dderny           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "vec3d.h"
 
-inline t_vec3d	vec3d_lerp(const t_vec3d a, const t_vec3d b, double t)
+inline t_vec3d	vec3d_lerp(const t_vec3d a, const t_vec3d b, float t)
 {
 	if (t < 0)
 		t = 0;
