+++ src/vec3d/normalize.c
@@ -6,7 +6,7 @@
 /*   By: dderny <dderny@student.42lyon.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/09 23:06:29 by dderny            #+#    #+#             */
-/*   Updated: 2025/05/11 14:43:58 by dderny           ###   ########.fr       */
+/*   Updated: 2025/05/24 05:41:54 by dderny           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 
 inline t_vec3d	vec3d_normalize(const t_vec3d v)
 {
-	double length = sqrt(v.x * v.x + v.y * v.y + v.z * v.z);
+	float length = sqrt(v.x * v.x + v.y * v.y + v.z * v.z);
 	if (length == 0)
 		return (v);
 	return ((t_vec3d){v.x / length, v.y / length, v.z / length});
