+++ src/vec3d/reflect.c
@@ -6,7 +6,7 @@
 /*   By: dderny <dderny@student.42lyon.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/09 02:19:02 by dderny            #+#    #+#             */
-/*   Updated: 2025/05/11 14:43:58 by dderny           ###   ########.fr       */
+/*   Updated: 2025/05/24 05:41:54 by dderny           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 inline t_vec3d	vec3d_reflect(const t_vec3d a, const t_vec3d normal)
 {
-	double		dot;
+	float		dot;
 
 	dot = vec3d_dot(a, normal);
 	return ((t_vec3d){
